pyproject.toml
[project]
name = "essay-grading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fuzzywuzzy==0.18.0",
    "gender-guesser>=0.4.0",
    "ipykernel>=6.29.5",
    "nltk>=3.9.1",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-levenshtein>=0.27.1",
    "pyyaml>=6.0.2",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.13.0",
    "rich>=14.0.0",
    "ruff>=0.11.7",
    "scipy>=1.15.2",
    "sentence-transformers>=4.1.0",
    "strsim==0.0.3",
    "textstat>=0.7.5",
]


[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]


[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true


[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "UP006",
    "ERA001",
    "UP007",
    "LOG015",
    "G004",
    "TRY300",
    "BLE001",
    "E731",
    "T201",
    "B019"
]

[tool.ruff.format]
quote-style = "double"
line-ending = "lf"